requirements.txt
streamlit
pandas
geopy
requests
beautifulsoup4
groq
python-dotenv
